+++ pydantic/dataclasses.py
@@ -44,7 +44,7 @@ def setattr_validate_assignment(self: 'DataclassType', name: str, value: Any) ->
         d.pop(name)
         value, error_ = self.__pydantic_model__.__fields__[name].validate(value, d, loc=name, cls=self.__class__)
         if error_:
-            raise ValidationError([error_])
+            raise ValidationError([error_], type(self))
 
     object.__setattr__(self, name, value)