+++ HISTORY.rst
@@ -3,6 +3,11 @@
 History
 -------
 
+
+v0.32 (unreleased)
+..................
+* add model name to ``ValidationError`` error message, #676 by @dmontagu
+
 v0.31.1 (2019-07-31)
 ....................
 * fix json generation for ``EnumError``, #697 by @dmontagu
@@ -20,6 +25,7 @@ v0.31 (2019-07-24)
 * add ``Config.keep_untouched`` for custom descriptors support, #679 by @MrMrRobat
 * use ``inspect.cleandoc`` internally to get model description, #657 by @tiangolo
 * add ``Color`` to schema generation, by @euri10
+* add documentation for Literal type, #651 by @dmontagu
 
 v0.30.1 (2019-07-15)
 ....................